Reinforcement_learning/AlphaZero-master/pyproject.toml
[project]
name = "alpha-zero"
version = "1.0.0"
description = "AlphaZero implementation for reinforcement learning through self-play"
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "python-chess>=1.999",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

